Refactor: ReqeustDto  삭제
Shceduler 를 통해 모든것이 자동화가 되어 더이상 RequestDto 가 필요하지 않게 되어 삭제조치 하였습니다.

related to: #59
