--- src/main/java/com/prgrms/ijuju/domain/article/dto/request/ArticleRequestDto.java
@@ -1,11 +1,0 @@
-package com.prgrms.ijuju.domain.article.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class ArticleRequestDto {
-    private String stockSymbol;
-    private String trendPrediction;
-    private String content;
-    private int duration;
-}
